feat(workflow): GA streamline-targeting issue Assigned To field
In streamline targeting we combined the Owned By field and the Assigned To fields into one and put the suggestions into the assigned to dropdown.

This cleanup removes the 2 components that are no longer used.
