+++ static/app/components/assigneeSelectorDropdown.tsx
@@ -142,10 +142,6 @@ export class AssigneeSelectorDropdown extends Component<
     }, undefined),
   ];
 
-  hasStreamlineTargeting() {
-    return this.props.organization.features.includes('streamline-targeting-context');
-  }
-
   handleMemberListUpdate = (members: User[]) => {
     if (members === this.state.memberList) {
       return;
@@ -252,29 +248,18 @@ export class AssigneeSelectorDropdown extends Component<
           <IconContainer>
             <UserAvatar user={member} size={24} />
           </IconContainer>
-          {this.hasStreamlineTargeting() ? (
-            <div>
-              <AssigneeLabel>
-                <Highlight text={inputValue}>
-                  {sessionUser.id === member.id
-                    ? `${member.name || member.email} ${t('(You)')}`
-                    : member.name || member.email}
-                </Highlight>
-              </AssigneeLabel>
-              {suggestedReason && (
-                <SuggestedAssigneeReason>{suggestedReason}</SuggestedAssigneeReason>
-              )}
-            </div>
-          ) : (
-            <Label>
+          <div>
+            <AssigneeLabel>
               <Highlight text={inputValue}>
                 {sessionUser.id === member.id
                   ? `${member.name || member.email} ${t('(You)')}`
                   : member.name || member.email}
               </Highlight>
-              {suggestedReason && <SuggestedReason> ({suggestedReason})</SuggestedReason>}
-            </Label>
-          )}
+            </AssigneeLabel>
+            {suggestedReason && (
+              <SuggestedAssigneeReason>{suggestedReason}</SuggestedAssigneeReason>
+            )}
+          </div>
         </MenuItemWrapper>
       ),
     };
@@ -301,21 +286,14 @@ export class AssigneeSelectorDropdown extends Component<
           <IconContainer>
             <TeamAvatar team={team} size={24} />
           </IconContainer>
-          {this.hasStreamlineTargeting() ? (
-            <div>
-              <AssigneeLabel>
-                <Highlight text={inputValue}>{display}</Highlight>
-              </AssigneeLabel>
-              {suggestedReason && (
-                <SuggestedAssigneeReason>{suggestedReason}</SuggestedAssigneeReason>
-              )}
-            </div>
-          ) : (
-            <Label>
+          <div>
+            <AssigneeLabel>
               <Highlight text={inputValue}>{display}</Highlight>
-              {suggestedReason && <SuggestedReason> ({suggestedReason})</SuggestedReason>}
-            </Label>
-          )}
+            </AssigneeLabel>
+            {suggestedReason && (
+              <SuggestedAssigneeReason>{suggestedReason}</SuggestedAssigneeReason>
+            )}
+          </div>
         </MenuItemWrapper>
       ),
     };
@@ -377,51 +355,22 @@ export class AssigneeSelectorDropdown extends Component<
     });
 
     // New version combines teams and users into one section
-    const dropdownItems: ItemsBeforeFilter = this.hasStreamlineTargeting()
-      ? [
-          {
-            label: this.renderDropdownGroupLabel(t('Everyone Else')),
-            hideGroupLabel: !filteredSuggestedAssignees.length,
-            id: 'everyone-else',
-            items: [...filteredSessionUser, ...filteredTeams, ...filteredMembers],
-          },
-        ]
-      : [
-          {
-            label: this.renderDropdownGroupLabel(t('Teams')),
-            id: 'team-header',
-            items: filteredTeams,
-          },
-          {
-            label: this.renderDropdownGroupLabel(t('People')),
-            id: 'members-header',
-            items: filteredMembers,
-          },
-        ];
+    const dropdownItems: ItemsBeforeFilter = [
+      {
+        label: this.renderDropdownGroupLabel(t('Everyone Else')),
+        hideGroupLabel: !filteredSuggestedAssignees.length,
+        id: 'everyone-else',
+        items: [...filteredSessionUser, ...filteredTeams, ...filteredMembers],
+      },
+    ];
 
     if (suggestedAssignees.length || filteredSessionUser.length) {
       // Add suggested assingees
-      if (this.hasStreamlineTargeting()) {
-        dropdownItems.unshift({
-          label: this.renderDropdownGroupLabel(t('Suggested Assignees')),
-          id: 'suggested-list',
-          items: filteredSuggestedAssignees,
-        });
-      } else {
-        dropdownItems.unshift(
-          // session user is first on dropdown
-          {
-            label: this.renderDropdownGroupLabel(t('Suggested')),
-            id: 'suggested-header',
-            items: filteredSessionUser,
-          },
-          {
-            hideGroupLabel: true,
-            id: 'suggested-list',
-            items: filteredSuggestedAssignees,
-          }
-        );
-      }
+      dropdownItems.unshift({
+        label: this.renderDropdownGroupLabel(t('Suggested Assignees')),
+        id: 'suggested-list',
+        items: filteredSuggestedAssignees,
+      });
     }
 
     return dropdownItems;
@@ -677,8 +626,4 @@ const GroupHeader = styled('div')`
   margin: ${space(1)} 0;
   color: ${p => p.theme.subText};
   text-align: left;
-`;
-
-const SuggestedReason = styled('span')`
-  color: ${p => p.theme.subText};
 `;
